Add `file` value source type
Also refactored config tests a bunch
