+++ .env-select.toml
@@ -1,8 +1,13 @@
 [applications.server.profiles.base]
 variables = {PROTOCOL = "https"}
+
 [applications.server.profiles.dev]
 extends = ["base"]
 variables = {SERVICE1 = "dev", SERVICE2 = "also-dev"}
+
 [applications.server.profiles.prd]
 extends = ["base"]
-variables = {SERVICE1 = "prd", SERVICE2 = "also-prd"}
+[applications.server.profiles.prd.variables]
+SERVICE1 = "prd"
+SERVICE2 = "also-prd"
+SERVICE3 = {type = "file", path = "creds.env", multiple = true}
