+++ creds.env
@@ -1,0 +1,3 @@
+# This is an example file
+USERNAME=username
+PASSWORD=password
